Compile template as a separate test
